refactor: moved EmployeeAuthority classes to separate package
