+++ src/main/java/com/pawelnu/projectmanager/config/DataInit.java
@@ -3,8 +3,8 @@ package com.pawelnu.projectmanager.config;
 import com.github.javafaker.Faker;
 import com.pawelnu.projectmanager.endpoints.authority.AuthorityEntity;
 import com.pawelnu.projectmanager.endpoints.authority.AuthorityRepository;
-import com.pawelnu.projectmanager.endpoints.authority.EmployeeAuthorityEntity;
-import com.pawelnu.projectmanager.endpoints.authority.EmployeeAuthorityRepository;
+import com.pawelnu.projectmanager.endpoints.authority.employee.EmployeeAuthorityEntity;
+import com.pawelnu.projectmanager.endpoints.authority.employee.EmployeeAuthorityRepository;
 import com.pawelnu.projectmanager.endpoints.company.CompanyEntity;
 import com.pawelnu.projectmanager.endpoints.company.CompanyRepository;
 import com.pawelnu.projectmanager.endpoints.company.address.CompanyAddressEntity;
