+++ src/main/java/com/pawelnu/projectmanager/endpoints/company/employee/EmployeeEntity.java
@@ -1,6 +1,6 @@
 package com.pawelnu.projectmanager.endpoints.company.employee;
 
-import com.pawelnu.projectmanager.endpoints.authority.EmployeeAuthorityEntity;
+import com.pawelnu.projectmanager.endpoints.authority.employee.EmployeeAuthorityEntity;
 import com.pawelnu.projectmanager.endpoints.category.value.CategoryValueEntity;
 import com.pawelnu.projectmanager.endpoints.company.CompanyEntity;
 import com.pawelnu.projectmanager.entity.Auditable;
